lynx: We don't support udfps touch handler
This commit fixes broken UDFPS on gs101/gs201/zuma. The
breakage is a combination of two changes: The UDFPS HAL fallback
on Android 16 switching from false to true, and the introduction
of I1b7eef5449f813deee23ebe9ff07d4d3b73ccf4b.

That change passes halHandlesDisplayTouches to
FingerprintSensorPropertiesInternal, which, when combined with the
A16 HAL fallback, causes the touch handler to be incorrectly
enabled. This went unnoticed on Android 15 because the HAL
defaulted to false and the fps_touch_handler_supported property
was not set.

Change-Id: If24475b5f14b7f02b90774be6acf20865285f1ab
Signed-off-by: AnierinB <anierin@evolution-x.org>